+++ RELEASENOTES.md
@@ -1,7 +1,8 @@
 # Release Notes
 
-### 3.8.3 (UNRELEASED)
+### 3.9.0 (UNRELEASED)
 * Added parent service body names and IDs for service body selection.
+* Saying links is now an option which is off by default.
 * User modal titling fixes.
 * Fixes for metrics reports and performance improvements.